a2a-communicator/requirements.txt
openai>=0.27.0
python-dotenv>=1.1.0
httpx>=0.28.1
pydantic>=2.7.2,<3.0.0
uvicorn>=0.23.1
anyio>=4.5 